Rename Java util logging javaagent package
