+++ instrumentation/java-util-logging/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/java/util/logging/JavaUtilLoggingInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.jul;
+package io.opentelemetry.javaagent.instrumentation.java.util.logging;
 
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.extendsClass;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
